requirements.txt
flask
langchain
google-generativeai
python-dotenv

